fix police check/ add item requirement check
Here we added the initial check server side for the item requirement. if the config is enforced players wont be able to access the black market with out having the item "blaclkmkticket" in their inventory. this will also remove it as a one time use token that players would have to grind for via your own bank robberies heists etc. 

weve also fixed the issue with the police being disallowed from accessing the UI.  initially this line:

   if not checkJobRestrictions(src) then
        TriggerClientEvent("nu-blackmarket:client:purchaseResult", src, false, "Access denied (job restricted)")
        return
    end

would only disallow the police to purchase items on duty. if police went OFF duty, they could still purchase items. With the fix implemented we now restrict police from accessing the ui at all. IF someone is able to bypaass this the secondary check would be the line shown above